requirements.txt
python-dotenv>=1.0.0,<2
langchain-core>=1.3.0,<2
langchain-community>=0.4.0,<0.5
langchain-text-splitters>=1.1.0,<2
langchain-classic>=1.0.0,<2
langchain-openai>=1.2.0,<2
chromadb>=0.5.0,<2
pypdf>=4.0.0,<7
sentence-transformers>=3.0.0,<6
torch>=2.0.0,<3
torchvision>=0.15.0,<2
streamlit>=1.40.0,<2
